backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
gunicorn==23.0.0
dj-database-url==2.2.0
